Add -Wextra to gcc in Makefile

+++ stub.c
@@ -8,7 +8,7 @@ void init_interpreter_list(void);
 
 int emacs_module_init(struct emacs_runtime* runtime) {
     // https://www.gnu.org/software/emacs/manual/html_node/elisp/Module-Initialization.html#index-emacs_005fmodule_005finit-1
-    if (runtime->size < sizeof (*runtime)) {
+    if ((long unsigned int) runtime->size < sizeof (*runtime)) {
         fprintf(stderr, "%s\n", \
                 "ERROR: emacs_module_init: emacspy was compiled for newer version of Emacs.");
         return 1;
